+++ Assets/MotionFramework/ThirdParty/YooAsset/Editor/AssetBundleCollector/AssetBundleCollectorWindow.cs
@@ -31,37 +31,35 @@ namespace YooAsset.Editor
 		private TextField _groupDescTxt;
 		private TextField _groupAssetTagsTxt;
 		private VisualElement _groupContainer;
+		private string _lastModifyGroup = string.Empty;
+
 
 		public void CreateGUI()
 		{
 			Undo.undoRedoPerformed -= RefreshWindow;
 			Undo.undoRedoPerformed += RefreshWindow;
 
-			VisualElement root = this.rootVisualElement;
-
-			_collectorTypeList = new List<string>()
-			{
-				$"{nameof(ECollectorType.MainAssetCollector)}",
-				$"{nameof(ECollectorType.StaticAssetCollector)}",
-				$"{nameof(ECollectorType.DependAssetCollector)}"
-			};
-			_addressRuleList = AssetBundleCollectorSettingData.GetAddressRuleNames();
-			_packRuleList = AssetBundleCollectorSettingData.GetPackRuleNames();
-			_filterRuleList = AssetBundleCollectorSettingData.GetFilterRuleNames();
-
-			// 加载布局文件
-			string rootPath = EditorTools.GetYooAssetSourcePath();
-			string uxml = $"{rootPath}/Editor/AssetBundleCollector/{nameof(AssetBundleCollectorWindow)}.uxml";
-			var visualAsset = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(uxml);
-			if (visualAsset == null)
-			{
-				Debug.LogError($"Not found {nameof(AssetBundleCollectorWindow)}.uxml : {uxml}");
-				return;
-			}
-			visualAsset.CloneTree(root);
-
 			try
 			{
+				_collectorTypeList = new List<string>()
+				{
+					$"{nameof(ECollectorType.MainAssetCollector)}",
+					$"{nameof(ECollectorType.StaticAssetCollector)}",
+					$"{nameof(ECollectorType.DependAssetCollector)}"
+				};
+				_addressRuleList = AssetBundleCollectorSettingData.GetAddressRuleNames();
+				_packRuleList = AssetBundleCollectorSettingData.GetPackRuleNames();
+				_filterRuleList = AssetBundleCollectorSettingData.GetFilterRuleNames();
+
+				VisualElement root = this.rootVisualElement;
+
+				// 加载布局文件
+				var visualAsset = EditorHelper.LoadWindowUXML<AssetBundleCollectorWindow>();
+				if (visualAsset == null)
+					return;
+				
+				visualAsset.CloneTree(root);
+
 				// 导入导出按钮
 				var exportBtn = root.Q<Button>("ExportButton");
 				exportBtn.clicked += ExportBtn_clicked;
@@ -166,6 +164,9 @@ namespace YooAsset.Editor
 		}
 		public void OnDestroy()
 		{
+			// 注意：清空所有撤销操作
+			Undo.ClearAll();
+
 			if (AssetBundleCollectorSettingData.IsDirty)
 				AssetBundleCollectorSettingData.SaveFile();
 		}
@@ -205,6 +206,16 @@ namespace YooAsset.Editor
 			_groupListView.ClearSelection();
 			_groupListView.itemsSource = AssetBundleCollectorSettingData.Setting.Groups;
 			_groupListView.Rebuild();
+
+			for (int index = 0; index < AssetBundleCollectorSettingData.Setting.Groups.Count; index++)
+			{
+				var group = AssetBundleCollectorSettingData.Setting.Groups[index];
+				if (group.GroupName == _lastModifyGroup)
+				{
+					_groupListView.selectedIndex = index;
+					break;
+				}
+			}
 		}
 		private VisualElement MakeGroupListViewItem()
 		{
@@ -238,7 +249,7 @@ namespace YooAsset.Editor
 		}
 		private void AddGroupBtn_clicked()
 		{
-			Undo.RecordObject(AssetBundleCollectorSettingData.Setting, "YooAsset AddGroup");
+			Undo.RecordObject(AssetBundleCollectorSettingData.Setting, "YooAsset.AssetBundleCollectorWindow AddGroup");
 			AssetBundleCollectorSettingData.CreateGroup("Default Group");
 			FillGroupViewData();
 		}
@@ -248,8 +259,7 @@ namespace YooAsset.Editor
 			if (selectGroup == null)
 				return;
 
-			Undo.RecordObject(AssetBundleCollectorSettingData.Setting, "YooAsset RemoveGroup");
-
+			Undo.RecordObject(AssetBundleCollectorSettingData.Setting, "YooAsset.AssetBundleCollectorWindow RemoveGroup");
 			AssetBundleCollectorSettingData.RemoveGroup(selectGroup);
 			FillGroupViewData();
 		}
@@ -264,6 +274,7 @@ namespace YooAsset.Editor
 				return;
 			}
 
+			_lastModifyGroup = selectGroup.GroupName;
 			_groupContainer.visible = true;
 			_groupNameTxt.SetValueWithoutNotify(selectGroup.GroupName);
 			_groupDescTxt.SetValueWithoutNotify(selectGroup.GroupDesc);
@@ -552,6 +563,7 @@ namespace YooAsset.Editor
 			if (selectGroup == null)
 				return;
 
+			Undo.RecordObject(AssetBundleCollectorSettingData.Setting, "YooAsset.AssetBundleCollectorWindow AddCollector");
 			AssetBundleCollectorSettingData.CreateCollector(selectGroup, string.Empty);
 			FillCollectorViewData();
 		}
@@ -562,6 +574,8 @@ namespace YooAsset.Editor
 				return;
 			if (selectCollector == null)
 				return;
+
+			Undo.RecordObject(AssetBundleCollectorSettingData.Setting, "YooAsset.AssetBundleCollectorWindow RemoveCollector");
 			AssetBundleCollectorSettingData.RemoveCollector(selectGroup, selectCollector);
 			FillCollectorViewData();
 		}