+++ Assets/MotionFramework/ThirdParty/YooAsset/Editor/AssetBundleDebugger/VisualViewers/DebuggerAssetListViewer.cs
@@ -9,7 +9,7 @@ using UnityEngine.UIElements;
 
 namespace YooAsset.Editor
 {
-	internal class AssetListDebuggerViewer
+	internal class DebuggerAssetListViewer
 	{
 		private VisualTreeAsset _visualAsset;
 		private TemplateContainer _root;
@@ -23,15 +23,11 @@ namespace YooAsset.Editor
 		/// </summary>
 		public void InitViewer()
 		{
-			// 加载布局文件
-			string rootPath = EditorTools.GetYooAssetSourcePath();
-			string uxml = $"{rootPath}/Editor/AssetBundleDebugger/VisualViewers/{nameof(AssetListDebuggerViewer)}.uxml";
-			_visualAsset = AssetDatabase.LoadAssetAtPath<VisualTreeAsset>(uxml);
+			// 加载布局文件		
+			_visualAsset = EditorHelper.LoadWindowUXML<DebuggerAssetListViewer>();
 			if (_visualAsset == null)
-			{
-				Debug.LogError($"Not found {nameof(AssetListDebuggerViewer)}.uxml : {uxml}");
 				return;
-			}
+			
 			_root = _visualAsset.CloneTree();
 			_root.style.flexGrow = 1f;