+++ Assets/MotionFramework/ThirdParty/YooAsset/Runtime/DownloadSystem/DownloadSystem.cs
@@ -15,7 +15,7 @@ namespace YooAsset
 		private static readonly Dictionary<string, DownloaderBase> _downloaderDic = new Dictionary<string, DownloaderBase>();
 		private static readonly List<string> _removeList = new List<string>(100);
 		private static readonly Dictionary<string, string> _cachedHashList = new Dictionary<string, string>(1000);
-		private static int _breakpointResumeFileSize;
+		private static int _breakpointResumeFileSize = int.MaxValue;
 
 		/// <summary>
 		/// 初始化
@@ -155,17 +155,24 @@ namespace YooAsset
 		}
 		public static bool CheckContentIntegrity(string filePath, long size, string crc)
 		{
-			if (File.Exists(filePath) == false)
-				return false;
+			try
+			{
+				if (File.Exists(filePath) == false)
+					return false;
 
-			// 先验证文件大小
-			long fileSize = FileUtility.GetFileSize(filePath);
-			if (fileSize != size)
-				return false;
+				// 先验证文件大小
+				long fileSize = FileUtility.GetFileSize(filePath);
+				if (fileSize != size)
+					return false;
 
-			// 再验证文件CRC
-			string fileCRC = HashUtility.FileCRC32(filePath);
-			return fileCRC == crc;
+				// 再验证文件CRC
+				string fileCRC = HashUtility.FileCRC32(filePath);
+				return fileCRC == crc;
+			}
+			catch(Exception)
+			{
+				return false;
+			}
 		}
 	}
 }