+++ Assets/MotionFramework/ThirdParty/YooAsset/Runtime/PatchSystem/AssetInfo.cs
@@ -1,8 +1,46 @@
-﻿
+﻿using System.IO;
+
 namespace YooAsset
 {
 	public class AssetInfo
 	{
+		private readonly PatchAsset _patchAsset;
+		private string _providerGUID;
+
+		/// <summary>
+		/// 资源提供者唯一标识符
+		/// </summary>
+		internal string ProviderGUID
+		{
+			get
+			{
+				if (string.IsNullOrEmpty(_providerGUID) == false)
+					return _providerGUID;
+
+				if (AssetType == null)
+					_providerGUID = $"{AssetPath}[null]";
+				else
+					_providerGUID = $"{AssetPath}[{AssetType.Name}]";
+				return _providerGUID;
+			}
+		}
+
+		/// <summary>
+		/// 身份是否无效
+		/// </summary>
+		internal bool IsInvalid
+		{
+			get
+			{
+				return _patchAsset == null;
+			}
+		}
+
+		/// <summary>
+		/// 错误信息
+		/// </summary>
+		internal string Error { private set; get; }
+
 		/// <summary>
 		/// 资源路径
 		/// </summary>
@@ -13,15 +51,37 @@ namespace YooAsset
 		/// </summary>
 		public System.Type AssetType { private set; get; }
 
-		public AssetInfo(string assetPath, System.Type assetType)
+
+		// 注意：这是一个内部类，严格限制外部创建。
+		private AssetInfo()
 		{
-			AssetPath = assetPath;
-			AssetType = assetType;
 		}
-		public AssetInfo(string assetPath)
+		internal AssetInfo(PatchAsset patchAsset, System.Type assetType)
 		{
-			AssetPath = assetPath;
+			if (patchAsset == null)
+				throw new System.Exception("Should never get here !");
+
+			_patchAsset = patchAsset;
+			AssetType = assetType;
+			AssetPath = patchAsset.AssetPath;
+			Error = string.Empty;
+		}
+		internal AssetInfo(PatchAsset patchAsset)
+		{
+			if (patchAsset == null)
+				throw new System.Exception("Should never get here !");
+
+			_patchAsset = patchAsset;
 			AssetType = null;
+			AssetPath = patchAsset.AssetPath;
+			Error = string.Empty;
+		}
+		internal AssetInfo(string error)
+		{
+			_patchAsset = null;
+			AssetType = null;
+			AssetPath = string.Empty;
+			Error = error;
 		}
 	}
 }